Support for inherited fields with annotations
Annotations are usually cached by doctrine so the same instance will
be retrived when asking for the annotation for the same inherited
method via two different child classes. Because we modify the
annotation instance for our own needs, we need to make sure that
this process is not made twice for the method.