Fetch entity from DB on demand only
This gives a chance to do something before the entity is fetched from DB.
Typically this could allow to set up some security logic (eg: via Doctrine
filters) in resolvers before trying to fetch an entity.

So this gives more flexibility while still keeping the out-of-the-box
support for getters with parameters.