Expose entities default values
This allow to have mutations with less mandatory fields and instead
rely on default values as defined in Doctrine model. That should make
the API easier to use and reduce the data transfered back and forth
on the network to the strict minimum.

This could also allow API users to query for default values to use to
pre-populate forms on client-side.