Allow multiple use of same operator on the same field
This was previously partially possible by using the condition level,
as shown in `multiple-filter-on-same-field.php`, but that meant that
condition level could not be used for somethine else.

With this commit we are able reserve condition level for high-level
"groups" of filter combined with `OR` logic and still express several
conditon with the same operator on the field level, as demonstrated
in `multiple-filter-on-same-field-with-alternative.php`.

So it's easier to express "this group of filters OR that group of filter".