Conditions on joins are grouped with non-join conditions
This is a major change in the filtering syntax.

First, the following terms were renamed as follow:

- `conditions` => `groups`
- `fields` => `conditions`
- `conditionLogic` => `groupLogic`
- `fieldsLogic` => `conditionsLogic`

This is hopefully less confusing and more semantically correct.
We now have ordered `groups` of unordered `conditions`.

Second, the `joins` was moved from top-level to within a `group`. And
the join itself cannot define a logical operator anymore. Instead the
`conditionsLogic` of a specific `group` is used for all conditions
in that group, recursively including all conditions coming from joins.

This allow to use different conditions on a same join across different
groups and still keep control of how they are combined with the `groupLogic`.
A concrete example of that lives in `alternative-groups-with-multiple-joins.php`.