Update DocBlockReader.php - Resolves issue on No type registered with key `\DateTime`
The error below is encountered when the getter method is like this:

```
/**
 * @return \DateTime
 */
public function getDatetimeUsageStart()
{
    return $this->datetime_usage_start;
}
```

ERROR: 
```
PHP Fatal error:  Uncaught GraphQL\Doctrine\Exception: No type registered with key `\DateTime`. Either correct the usage, or register it in your custom types container when instantiating `GraphQL\Doctrine\Types`.
```

So the solution is to remove the leading backward-slash.