Lists always have non-null elements
Because in the vast majority of cases a list of something do not allow
null. This is especially true for Doctrine collection of entities returned
by getter, but also for most inputs used in setter.

If a list of mixed null and non-null elements is needed then a custom
type should be defined via annotation.