Sorting with NULL as highest in a single sorting
Previously we had to specify two sortings to sort with NULL as highest.
Now only one sorting is required, making usage of the API more intuitive,
and avoid duplicating `order` value unnecessarily.

This is not a breaking change because double sorting will still give
the same result, but it is not necessary anymore and should be
simplified as follow:

Before:

```graphql
query {
  posts(sorting: [
    {field: title, order: ASC, nullAsHighest: true},
    {field: title, order: ASC}
  ]) {
        title
  }
}
```

After:

```graphql
query {
  posts(sorting: [
    {field: title, order: ASC, nullAsHighest: true}
  ]) {
        title
  }
}
```