Initial draft of PHP 8 attribute support.
This should allow users to use this with the Doctrine AttributeDriver for metadata, instead of using the AnnotationDriver. This requires the that all `GraphQL\Doctrine\Annotation`s are written as annotations as well, though.
The alternative would be to use two readers, one that reads annotations and one that reads attributes, and fall back to one of the if the other does not yield a result.