Clarify type syntax #9954
`API\Field` and `API\Input` must always specify a FQCN if referencing
models. Otherwise, `ecodev/graphql-doctrine` is not able to find the
corresponding GraphQL type. So from `'User'`, it cannot guess we really
mean `'Application\Model\User'`.

The easiest solution is to use `::class` notation as much as possible,
even for non-models:

```diff
- #[API\Field(type: 'Group')]
+ #[API\Field(type: Group::class)]
```

```diff
- #[API\Input(type: 'Login')]
+ #[API\Input(type: LoginType::class)]
```

And keep string literals only when we must specify a nullable or list
type, such as:

```php
#[API\Field(type: '?Login')]
```

or

```php
#[API\Input(type: '?TagID[]')]
```