BREAKING New `TypesInterface::loadType()` #10525
This should be used to declare typeLoader in `GraphQL\Type\Schema` with
something similar to:

```php
$types = new Types(...);
$schema = new GraphQL\Type\Schema([
    'typeLoader' => fn (string $name) => $types->loadType($name, 'Application\Model') ?? $types->loadType($name, 'OtherApplication\Model')
    // ...
]);
```

While this method could technically replace of uses of dedicated `get*()
`methods, we suggest to only use `loadType` with the `typeLoader`.
Because dedicated `get*()` methods are easier to use, and provide
stronger typing.

This is a breaking change because of the new method on `TypesInterface`,
but pre-existing behavior remains unchanged.